requirements.txt
openai>=1.30.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
playwright>=1.44.0
httpx>=0.27.0
reportlab>=4.1.0
